+++ fx.go
@@ -16,7 +16,7 @@ import (
 	"github.com/urfave/cli"
 )
 
-const version = "0.8.72"
+const version = "0.8.73"
 
 func init() {
 	go checkForUpdate()
